refactor: 104-11.3.2 Moved User Create and user Me endpoint to user endpoints instead of auth endpoint. Moved tests so they are correct placed and changed routing logic to access the moved endpoints correctly.
