+++ app/api/v1/endpoints/auth.py
@@ -1,43 +1,14 @@
 from typing import Annotated
 from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.orm import Session
-from app.api.v1.schemas.user_schema import UserCreate, UserRead
 from app.api.v1.schemas.auth_schema import LoginRequest, Token
-from app.services import auth_service, user_service
-from app.dependencies import get_db, get_current_user
-from app.models.user_model import User
+from app.services import auth_service
+from app.dependencies import get_db
+
 
 router = APIRouter()
 
 DbSession = Annotated[Session, Depends(get_db)]
-
-
-@router.post(
-    "/register",
-    response_model=UserRead,
-    status_code=status.HTTP_201_CREATED,
-    summary="Register a new user",
-)
-async def register_user(user: UserCreate, db: DbSession):
-    """
-    Registers a new user with a username, password, and role.
-
-    Args:
-        user (UserCreate): Pydantic model containing username, password, and role.
-        db (Session): Database session dependency.
-
-    Returns:
-        UserRead: The newly created user object.
-
-    Raises:
-        HTTPException 400: If the username is already taken.
-
-    Responses:
-        201 Created: Successfully created user.
-        400 Bad Request: Username already exists.
-    """
-    new_user = user_service.create_user(db, user)
-    return new_user
 
 
 @router.post(
@@ -72,25 +43,3 @@ async def login_for_access_token(login_data: LoginRequest, db: DbSession):
         )
     access_token = auth_service.create_access_token_for_user(user)
     return {"access_token": access_token, "token_type": "bearer"}
-
-
-@router.get(
-    "/me",
-    response_model=UserRead,
-    summary="Get your own user",
-)
-async def fetch_current_user(current_user: Annotated[User, Depends(get_current_user)]):
-    """
-    Fetch the currently authenticated user's information.
-
-    Args:
-        current_user (User): Injected via JWT authentication dependency.
-
-    Returns:
-        UserRead: The currently authenticated user's details.
-
-    Responses:
-        200 OK: Successfully retrieved user data.
-        401 Unauthorized: Invalid or missing JWT token.
-    """
-    return current_user
