+++ tests/integration/test_user_endpoints.py
@@ -9,6 +9,13 @@ client = TestClient(app)
 # Fixtures
 # -----------------------------
 @pytest.fixture
+def new_user_data():
+    """
+    Purpose: Provides a new user payload for registration and login tests.
+    """
+    return {"username": "testuser", "password": "secret123", "role": "customer"}
+
+@pytest.fixture
 def auth_headers(client):
     """
     Purpose: Provide valid authentication headers for API requests.
@@ -16,7 +23,7 @@ def auth_headers(client):
     unique_username = f"apiuser_{uuid4()}"
     user_data = {"username": unique_username, "password": "1234", "role": "customer"}
 
-    client.post("/api/v1/auth/register", json=user_data)
+    client.post("/api/v1/users/register", json=user_data)
     login_resp = client.post("/api/v1/auth/login", json={"username": unique_username, "password": "1234"})
     token = login_resp.json()["access_token"]
 
@@ -26,13 +33,35 @@ def auth_headers(client):
 # -----------------------------
 # API Endpoint Tests
 # -----------------------------
+def test_register_user_success(new_user_data):
+    """
+    Purpose: Test that a new user can register successfully.
+    Scenario: POST /auth/register with valid username, password, and role.
+    Expected: HTTP 201 Created, response contains correct username and role.
+    """
+    response = client.post("/api/v1/users/register", json=new_user_data)
+    assert response.status_code == 201
+    data = response.json()
+    assert data["username"] == new_user_data["username"]
+    assert data["role"] == new_user_data["role"]
+
+def test_register_user_duplicate(new_user_data):
+    """
+    Purpose: Test registration with duplicate username fails.
+    Scenario: POST /auth/register twice with the same username.
+    Expected: HTTP 400 Bad Request on the second attempt.
+    """
+    client.post("/api/v1/users/register", json=new_user_data)
+    response = client.post("/api/v1/users/register", json=new_user_data)
+    assert response.status_code == 400
+
 def test_fetch_current_user(client, auth_headers):
     """
     Purpose: Verify fetching the currently authenticated user's info.
     Scenario: Send GET request to /api/v1/auth/me with valid token headers.
     Expected: Returns 200 OK with username in response JSON.
     """
-    response = client.get("/api/v1/auth/me", headers=auth_headers)
+    response = client.get("/api/v1/users/me", headers=auth_headers)
     assert response.status_code == 200
     data = response.json()
     assert "username" in data
@@ -62,7 +91,7 @@ def test_admin_crud_users_flow(client):
     """
     # Create admin
     admin_username = f"admin_{uuid4()}"
-    client.post("/api/v1/auth/register", json={"username": admin_username, "password": "a", "role": "admin"})
+    client.post("/api/v1/users/register", json={"username": admin_username, "password": "a", "role": "admin"})
 
     login_resp = client.post("/api/v1/auth/login", json={"username": admin_username, "password": "a"})
     admin_token = login_resp.json()["access_token"]
@@ -70,7 +99,7 @@ def test_admin_crud_users_flow(client):
 
     # Create normal user
     user_username = f"user_{uuid4()}"
-    create_resp = client.post("/api/v1/auth/register", json={"username": user_username, "password": "b", "role": "customer"})
+    create_resp = client.post("/api/v1/users/register", json={"username": user_username, "password": "b", "role": "customer"})
     user_id = create_resp.json()["id"]
 
     # GET by id
